K_max_and_min.cpp

+++ CodeForces/ICPC Assiut University Community/Sheet#1 Data type-Conditions/K_max_and_min.cpp
@@ -4,7 +4,7 @@ using namespace std;
 #define mod 1000000007
 void solve()
 {
-    long long int a, b, c, max, min;
+    ll a, b, c, max, min;
 
     cin >> a >> b >> c;
 
@@ -62,16 +62,19 @@ void solve()
     cout << "" << min << " "
          << "" << max << endl;
 }
-int main() {
-    ios_base::sync_with_stdio(false);cin.tie(NULL);cout.tie(NULL);
+int main()
+{
+    ios_base::sync_with_stdio(false);
+    cin.tie(NULL);
+    cout.tie(NULL);
 #ifndef ONLINE_JUDGE
     freopen("input.txt", "r", stdin);
     freopen("output.txt", "w", stdout);
 #endif
 
-    ll test=1;
-//cin>>test;
-    while(test--)
+    ll test = 1;
+    // cin>>test;
+    while (test--)
     {
         solve();
     }
